Revision: 201021 Kit: 2010123

+++ appinstaller/AppinstUi/sisxsilentinstallindicatorplugin/inc/sisxsilentinstallindicator.h
@@ -58,7 +58,7 @@ private:
     
 private:
     
-    float mUpdateValue;
+    int mUpdateValue;
     bool mIsInstallProcess;
     
 };
